refactor(proxy-gateway): split Responses request normalisation out of the controller
Group 3 continued, and the last of the behavior-preserving splits in it.

`OpenAIController` carried about 430 lines of pure normalisation turning the
Responses request shape into the chat request the gateway already serves, plus
the small guards around it. The cluster was closed: every `this.` call inside it
landed on another member of the same set, and `buildResponsesChatRequest` alone
was 171 lines.

It now lives in `modules/openai/responses/openai-responses-request.ts` as free
functions. The controller keeps the routes, the guards and the orchestration.

`ResponsesRequestBody` moved with it, because the module needs the type and
importing it back from the controller would have made the two files circular.
The controller re-exports it, so `src/server/main.ts` keeps its existing import.

Nothing outside the controller reached any of these, checked by name and by
string so reflective access would show up too.

  openai.controller.ts   1112 -> 682 lines
  responses/openai-responses-request.ts   new, 452 lines

Group 3 in total: `openai.service.ts` 1394 -> 888, `openai.controller.ts`
1112 -> 682, three new modules under the OpenAI surface. No protocol boundary was
moved and nothing went into shared code, since all of it is OpenAI-specific.

Verified: `tsc --noEmit` clean, `eslint .` 0 errors with the same 8 warnings as
the base, `prettier --check .` clean, `vitest run` 818 passed and 4 skipped with
no failures.
