+++ src/modules/proxy-gateway/server/modules/openai/openai.controller.ts
@@ -11,18 +11,15 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { FastifyReply, FastifyRequest } from 'fastify';
-import { isEmpty, isNil, isPlainObject, isString } from 'lodash-es';
+import { isEmpty, isString } from 'lodash-es';
 import { Observable } from 'rxjs';
 import {
   OpenAIChatRequest,
-  OpenAIToolCall,
   OpenAIChatResponse,
   OpenAIContentPart,
   GeminiRequest,
   GeminiResponse,
 } from '@/modules/proxy-gateway/server/common/interfaces/request-interfaces';
-import { toCustomToolArguments } from '@/modules/proxy-gateway/antigravity/CustomToolCall';
-import { ApplyPatchFailureCompactor } from '@/modules/proxy-gateway/antigravity/ApplyPatchFailureCompaction';
 import { toOpenAIResponsesResponse } from '@/modules/proxy-gateway/antigravity/OpenAIResponsesResponseMapper';
 import {
   mergeOpenAIResponsesInputItems,
@@ -48,28 +45,21 @@ import { safeStringifyPacket } from '@/shared/security/sensitiveDataMasking';
 import { BaseProxyController } from '@/modules/proxy-gateway/server/common/base-proxy.controller';
 import { ProxyGuard } from '@/modules/proxy-gateway/server/guards/proxy.guard';
 import { OpenAIService } from './openai.service';
+export type { ResponsesRequestBody } from './responses/openai-responses-request';
+import {
+  asString,
+  buildResponsesChatRequest,
+  type ResponsesRequestBody,
+  extractCompletedResponsesEvent,
+  normalizeResponsesInputItems,
+  resolveImageUrl,
+  resolveInlineData,
+  toRecord,
+} from './responses/openai-responses-request';
 
 export const IMAGE_QUOTA_REFRESH = Symbol('IMAGE_QUOTA_REFRESH');
 
 export type ImageQuotaRefresh = () => Promise<void>;
-
-export interface ResponsesRequestBody {
-  model?: string;
-  instructions?: string;
-  input?: unknown;
-  metadata?: Record<string, unknown>;
-  previous_response_id?: string;
-  tools?: OpenAIChatRequest['tools'];
-  max_output_tokens?: number;
-  temperature?: number;
-  top_p?: number;
-  presence_penalty?: number;
-  frequency_penalty?: number;
-  seed?: number;
-  tool_choice?: OpenAIChatRequest['tool_choice'];
-  stream?: boolean;
-  user?: string;
-}
 
 export interface PreparedResponsesRequest {
   request: OpenAIChatRequest;
@@ -297,7 +287,7 @@ export class OpenAIController extends BaseProxyController {
       return;
     }
 
-    const inlineAudio = this.resolveInlineData(body.file ?? body.audio, 'audio/mpeg');
+    const inlineAudio = resolveInlineData(body.file ?? body.audio, 'audio/mpeg');
     if (!inlineAudio) {
       res.status(HttpStatus.BAD_REQUEST).send({
         error: {
@@ -389,7 +379,7 @@ export class OpenAIController extends BaseProxyController {
   }
 
   public prepareResponsesRequest(body: ResponsesRequestBody): PreparedResponsesRequest | null {
-    const currentInputItems = this.normalizeResponsesInputItems(body.input);
+    const currentInputItems = normalizeResponsesInputItems(body.input);
     const previousSession = body.previous_response_id
       ? OpenAIResponsesSessionStore.get(body.previous_response_id)
       : null;
@@ -405,7 +395,7 @@ export class OpenAIController extends BaseProxyController {
     const model = body.model ?? previousSession?.model ?? 'gemini-3-flash';
     const instructions = body.instructions ?? previousSession?.instructions;
     const tools = body.tools ?? previousSession?.tools;
-    const request = this.buildResponsesChatRequest({
+    const request = buildResponsesChatRequest({
       ...body,
       input: inputItems,
       instructions,
@@ -424,24 +414,6 @@ export class OpenAIController extends BaseProxyController {
     };
   }
 
-  private normalizeResponsesInputItems(input: unknown): unknown[] {
-    if (Array.isArray(input)) {
-      return input;
-    }
-    if (isNil(input)) {
-      return [];
-    }
-
-    const content = isString(input) ? input : this.normalizeResponsesInput(input);
-    return [
-      {
-        type: 'message',
-        role: 'user',
-        content: [{ type: 'input_text', text: content }],
-      },
-    ];
-  }
-
   private cacheResponsesStream(
     stream: Observable<unknown>,
     session: OpenAIResponsesSession,
@@ -449,7 +421,7 @@ export class OpenAIController extends BaseProxyController {
     return new Observable<unknown>((subscriber) => {
       const subscription = stream.subscribe({
         next: (event) => {
-          this.saveResponsesSession(this.extractCompletedResponsesEvent(event), session);
+          this.saveResponsesSession(extractCompletedResponsesEvent(event), session);
           subscriber.next(event);
         },
         error: (error: unknown) => subscriber.error(error),
@@ -460,27 +432,9 @@ export class OpenAIController extends BaseProxyController {
     });
   }
 
-  private extractCompletedResponsesEvent(event: unknown): unknown | null {
-    if (!isString(event)) {
-      return null;
-    }
-
-    const dataLine = event.split(/\r?\n/).find((line) => line.startsWith('data:'));
-    if (!dataLine) {
-      return null;
-    }
-
-    try {
-      const parsed = this.toRecord(JSON.parse(dataLine.slice('data:'.length).trimStart()));
-      return parsed?.type === 'response.completed' ? (parsed.response ?? null) : null;
-    } catch {
-      return null;
-    }
-  }
-
   private saveResponsesSession(response: unknown, session: OpenAIResponsesSession): void {
-    const responseRecord = this.toRecord(response);
-    const responseId = this.asString(responseRecord?.id);
+    const responseRecord = toRecord(response);
+    const responseId = asString(responseRecord?.id);
     const output = responseRecord?.output;
     if (!responseId || !Array.isArray(output)) {
       return;
@@ -492,341 +446,13 @@ export class OpenAIController extends BaseProxyController {
     });
   }
 
-  private normalizeResponsesInput(input: unknown): string {
-    if (isString(input)) {
-      return input;
-    }
-
-    if (Array.isArray(input)) {
-      return input
-        .map((item) => {
-          if (isString(item)) {
-            return item;
-          }
-          const itemRecord = this.toRecord(item);
-          const content = this.asString(itemRecord?.content);
-          if (content) {
-            return content;
-          }
-          return JSON.stringify(item);
-        })
-        .join('\n');
-    }
-
-    if (isNil(input)) {
-      return '';
-    }
-
-    return JSON.stringify(input);
-  }
-
-  private buildResponsesChatRequest(body: ResponsesRequestBody): OpenAIChatRequest {
-    const messages: OpenAIChatRequest['messages'] = [];
-    if (isString(body.instructions) && !isEmpty(body.instructions.trim())) {
-      messages.push({
-        role: 'system',
-        content: body.instructions,
-      });
-    }
-
-    const callIdToToolName = new Map<string, string>();
-    const incompleteCustomCallIds = new Set<string>();
-    const applyPatchFailureCompactor = new ApplyPatchFailureCompactor();
-    const inputItems = Array.isArray(body.input) ? body.input : null;
-
-    if (inputItems) {
-      for (const item of inputItems) {
-        const itemObj = this.toRecord(item);
-        if (!itemObj) {
-          continue;
-        }
-
-        const type = this.asString(itemObj.type);
-        if (!type) {
-          continue;
-        }
-
-        if (
-          type === 'function_call' ||
-          type === 'local_shell_call' ||
-          type === 'web_search_call' ||
-          type === 'custom_tool_call'
-        ) {
-          const callId =
-            this.asString(itemObj.call_id) ?? this.asString(itemObj.id) ?? `call_${Date.now()}`;
-          if (
-            type === 'custom_tool_call' &&
-            this.asString(itemObj.status)?.toLowerCase() === 'incomplete'
-          ) {
-            incompleteCustomCallIds.add(callId);
-            continue;
-          }
-
-          const toolName =
-            type === 'local_shell_call'
-              ? 'shell'
-              : type === 'web_search_call'
-                ? 'builtin_web_search'
-                : (this.asString(itemObj.name) ?? 'unknown');
-          callIdToToolName.set(callId, toolName);
-        }
-      }
-
-      for (const item of inputItems) {
-        const itemObj = this.toRecord(item);
-        if (!itemObj) {
-          continue;
-        }
-
-        const type = this.asString(itemObj.type);
-        if (!type) {
-          continue;
-        }
-
-        if (type === 'message') {
-          const role = this.asString(itemObj.role) ?? 'user';
-          const content = this.normalizeResponsesMessageContent(itemObj.content);
-          messages.push({ role, content });
-          continue;
-        }
-
-        if (
-          type === 'function_call' ||
-          type === 'local_shell_call' ||
-          type === 'web_search_call' ||
-          type === 'custom_tool_call'
-        ) {
-          const callId =
-            this.asString(itemObj.call_id) ?? this.asString(itemObj.id) ?? `call_${Date.now()}`;
-          if (incompleteCustomCallIds.has(callId)) {
-            continue;
-          }
-
-          const toolName = callIdToToolName.get(callId) ?? 'unknown';
-          const customInput =
-            type === 'custom_tool_call' ? (this.asString(itemObj.input) ?? '') : undefined;
-          const args =
-            customInput === undefined
-              ? this.resolveToolArguments(type, itemObj)
-              : toCustomToolArguments(toolName, customInput);
-          const toolCall: OpenAIToolCall = {
-            id: callId,
-            type: 'function',
-            function: {
-              name: toolName,
-              arguments: JSON.stringify(args),
-            },
-          };
-          if (customInput !== undefined) {
-            toolCall.custom_input = customInput;
-          }
-          messages.push({
-            role: 'assistant',
-            content: '',
-            tool_calls: [toolCall],
-          });
-          continue;
-        }
-
-        if (type === 'function_call_output' || type === 'custom_tool_call_output') {
-          const callId = this.asString(itemObj.call_id) ?? this.asString(itemObj.id) ?? 'unknown';
-          if (incompleteCustomCallIds.has(callId)) {
-            continue;
-          }
-          if (type === 'custom_tool_call_output' && !callIdToToolName.has(callId)) {
-            continue;
-          }
-
-          const toolName = callIdToToolName.get(callId) ?? 'unknown';
-          const normalizedOutput = this.normalizeResponsesOutput(itemObj.output);
-          const output =
-            toolName === 'apply_patch'
-              ? applyPatchFailureCompactor.compact(normalizedOutput)
-              : normalizedOutput;
-          messages.push({
-            role: 'tool',
-            tool_call_id: callId,
-            name: toolName,
-            content: output,
-          });
-          continue;
-        }
-      }
-    } else if (isString(body.input)) {
-      messages.push({
-        role: 'user',
-        content: body.input,
-      });
-    } else if (!isNil(body.input)) {
-      messages.push({
-        role: 'user',
-        content: this.normalizeResponsesInput(body.input),
-      });
-    }
-
-    if (messages.length === 0) {
-      messages.push({
-        role: 'user',
-        content: '',
-      });
-    }
-
-    return {
-      model: body.model ?? 'gemini-3-flash',
-      messages,
-      tools: body.tools,
-      max_tokens: body.max_output_tokens,
-      temperature: body.temperature,
-      top_p: body.top_p,
-      presence_penalty: body.presence_penalty,
-      frequency_penalty: body.frequency_penalty,
-      seed: body.seed,
-      tool_choice: body.tool_choice,
-      stream: body.stream,
-      extra: {
-        ...(body.metadata ?? {}),
-        previous_response_id: body.previous_response_id,
-        user_id: body.user,
-      },
-    };
-  }
-
-  private normalizeResponsesMessageContent(content: unknown): string | OpenAIContentPart[] {
-    if (isString(content)) {
-      return content;
-    }
-
-    if (!Array.isArray(content)) {
-      return this.normalizeResponsesInput(content);
-    }
-
-    const textParts: string[] = [];
-    const imageParts: OpenAIContentPart[] = [];
-
-    for (const item of content) {
-      const block = this.toRecord(item);
-      if (!block) {
-        continue;
-      }
-
-      const blockType = this.asString(block.type);
-      if (blockType === 'input_text' || blockType === 'text' || blockType === 'output_text') {
-        const text = this.asString(block.text);
-        if (text) {
-          textParts.push(text);
-        }
-        continue;
-      }
-
-      if (blockType === 'input_image' || blockType === 'image_url') {
-        const imageUrl = this.resolveImageUrl(block);
-        if (imageUrl) {
-          imageParts.push({
-            type: 'image_url',
-            image_url: {
-              url: imageUrl,
-            },
-          });
-        }
-      }
-    }
-
-    if (imageParts.length === 0) {
-      return textParts.join('\n');
-    }
-
-    const merged: OpenAIContentPart[] = [];
-    if (textParts.length > 0) {
-      merged.push({
-        type: 'text',
-        text: textParts.join('\n'),
-      });
-    }
-    merged.push(...imageParts);
-    return merged;
-  }
-
-  private resolveToolArguments(
-    type: string,
-    item: Record<string, unknown>,
-  ): Record<string, unknown> {
-    if (type === 'local_shell_call') {
-      const action = this.toRecord(item.action);
-      const exec = action ? this.toRecord(action.exec) : null;
-      const command = this.asString(exec?.command);
-      return {
-        command: command ? [command] : [],
-      };
-    }
-
-    if (type === 'web_search_call') {
-      const action = this.toRecord(item.action);
-      return {
-        query: this.asString(action?.query) ?? '',
-      };
-    }
-
-    const raw = item.arguments;
-    if (isString(raw)) {
-      try {
-        const parsed = JSON.parse(raw);
-        const parsedRecord = this.toRecord(parsed);
-        if (parsedRecord) {
-          return parsedRecord;
-        }
-        return {
-          value: parsed,
-        };
-      } catch {
-        return {
-          raw,
-        };
-      }
-    }
-
-    const rawRecord = this.toRecord(raw);
-    if (rawRecord) {
-      return rawRecord;
-    }
-
-    return {};
-  }
-
-  private normalizeResponsesOutput(output: unknown): string {
-    if (isString(output)) {
-      return output;
-    }
-    const outputRecord = this.toRecord(output);
-    const content = this.asString(outputRecord?.content);
-    if (content) {
-      return content;
-    }
-    if (isNil(output)) {
-      return '';
-    }
-    return JSON.stringify(output);
-  }
-
-  private resolveImageUrl(block: Record<string, unknown>): string | null {
-    const raw = block.image_url;
-    if (isString(raw)) {
-      return raw;
-    }
-    const rawRecord = this.toRecord(raw);
-    const url = this.asString(rawRecord?.url);
-    if (url) {
-      return url;
-    }
-    return null;
-  }
-
   private collectImageContentParts(
     entries: Array<string | { data?: string; mimeType?: string } | undefined>,
     defaultMimeType: string,
   ): OpenAIContentPart[] {
     const parts: OpenAIContentPart[] = [];
     for (const entry of entries) {
-      const inlineData = this.resolveInlineData(entry, defaultMimeType);
+      const inlineData = resolveInlineData(entry, defaultMimeType);
       if (!inlineData) {
         continue;
       }
@@ -838,62 +464,6 @@ export class OpenAIController extends BaseProxyController {
       });
     }
     return parts;
-  }
-
-  private resolveInlineData(
-    input: unknown,
-    defaultMimeType: string,
-  ): {
-    mimeType: string;
-    data: string;
-  } | null {
-    if (!input) {
-      return null;
-    }
-
-    if (isString(input)) {
-      const dataUri = input.match(/^data:(?<mime>[^;]+);base64,(?<data>[A-Za-z0-9+/=]+)$/);
-      if (dataUri?.groups?.mime && dataUri.groups.data) {
-        return {
-          mimeType: dataUri.groups.mime,
-          data: dataUri.groups.data,
-        };
-      }
-
-      const cleaned = input.replace(/\s+/g, '');
-      if (cleaned.length > 0) {
-        return {
-          mimeType: defaultMimeType,
-          data: cleaned,
-        };
-      }
-      return null;
-    }
-
-    const inputRecord = this.toRecord(input);
-    if (inputRecord) {
-      const data = this.asString(inputRecord.data);
-      if (!data) {
-        return null;
-      }
-      return {
-        mimeType: this.asString(inputRecord.mimeType) ?? defaultMimeType,
-        data,
-      };
-    }
-
-    return null;
-  }
-
-  private toRecord(value: unknown): Record<string, unknown> | null {
-    if (!isPlainObject(value)) {
-      return null;
-    }
-    return value as Record<string, unknown>;
-  }
-
-  private asString(value: unknown): string | null {
-    return isString(value) ? value : null;
   }
 
   private async sendOpenAIImageGenerationResponse(
@@ -1067,8 +637,8 @@ export class OpenAIController extends BaseProxyController {
           textParts.push(block.text);
         }
         if (block.type === 'image_url') {
-          const imageUrl = this.resolveImageUrl(block as unknown as Record<string, unknown>);
-          const inlineData = this.resolveInlineData(imageUrl, 'image/png');
+          const imageUrl = resolveImageUrl(block as unknown as Record<string, unknown>);
+          const inlineData = resolveInlineData(imageUrl, 'image/png');
           if (inlineData) {
             parts.push({
               inlineData: {
